+++ src/main/java/nro/services/func/RadaService.java
@@ -125,8 +125,8 @@ public class RadaService {
                 List<ItemOption> options = cardT.getOptions();
                 ds.writeByte(options.size());
                 for (ItemOption option : options) {
-                    ds.writeByte(option.optionTemplate.id);
-                    ds.writeShort(option.param);
+                    ds.writeShort(option.optionTemplate.id);
+                    ds.writeInt(option.param);
                     ds.writeByte(option.activeCard);
                 }
             }
