+++ src/main/java/nro/services/func/ShopService.java
@@ -294,8 +294,8 @@ public class ShopService {
                         msg.writer().writeLong(itemShop.getPowerRequire());
                         msg.writer().writeByte(itemShop.options.size());
                         for (ItemOption option : itemShop.options) {
-                            msg.writer().writeByte(option.optionTemplate.id);
-                            msg.writer().writeShort(option.param);
+                            msg.writer().writeShort(option.optionTemplate.id);
+                            msg.writer().writeInt(option.param);
                         }
                         msg.writer().writeByte(itemShop.isNew ? 1 : 0);
                         msg.writer().writeByte(0);
@@ -326,8 +326,8 @@ public class ShopService {
                         msg.writer().writeInt(itemShop.gem);
                         msg.writer().writeByte(itemShop.options.size());
                         for (ItemOption option : itemShop.options) {
-                            msg.writer().writeByte(option.optionTemplate.id);
-                            msg.writer().writeShort(option.param);
+                            msg.writer().writeShort(option.optionTemplate.id);
+                            msg.writer().writeInt(option.param);
                         }
                         msg.writer().writeByte(itemShop.isNew ? 1 : 0);
                         CaiTrang caiTrang = Manager.getCaiTrangByItemId(itemShop.temp.id);
@@ -366,8 +366,8 @@ public class ShopService {
                         msg.writer().writeInt(itemShop.costSpec);
                         msg.writer().writeByte(itemShop.options.size());
                         for (ItemOption option : itemShop.options) {
-                            msg.writer().writeByte(option.optionTemplate.id);
-                            msg.writer().writeShort(option.param);
+                            msg.writer().writeShort(option.optionTemplate.id);
+                            msg.writer().writeInt(option.param);
                         }
                         msg.writer().writeByte(itemShop.isNew ? 1 : 0);
                         CaiTrang caiTrang = Manager.getCaiTrangByItemId(itemShop.temp.id);
@@ -899,8 +899,8 @@ public class ShopService {
                 List<ItemOption> itemOptions = item.getDisplayOptions();
                 msg.writer().writeByte(itemOptions.size());
                 for (ItemOption io : itemOptions) {
-                    msg.writer().writeByte(io.optionTemplate.id);
-                    msg.writer().writeShort(io.param);
+                    msg.writer().writeShort(io.optionTemplate.id);
+                    msg.writer().writeInt(io.param);
                 }
                 msg.writer().writeByte(1);
                 CaiTrang ct = Manager.getCaiTrangByItemId(item.template.id);
@@ -987,8 +987,8 @@ public class ShopService {
                 List<ItemOption> itemOptions = item.getDisplayOptions();
                 msg.writer().writeByte(itemOptions.size() + 1);
                 for (ItemOption io : itemOptions) {
-                    msg.writer().writeByte(io.optionTemplate.id);
-                    msg.writer().writeShort(io.param);
+                    msg.writer().writeShort(io.optionTemplate.id);
+                    msg.writer().writeInt(io.param);
                 }
                 //số lượng
                 msg.writer().writeByte(31);
