+++ src/main/java/nro/services/func/Trade.java
@@ -379,8 +379,8 @@ public class Trade {
                     List<ItemOption> itemOptions = item.getDisplayOptions();
                     ds.writeByte(itemOptions.size());
                     for (ItemOption io : itemOptions) {
-                        ds.writeByte(io.optionTemplate.id);
-                        ds.writeShort(io.param);
+                        ds.writeShort(io.optionTemplate.id);
+                        ds.writeInt(io.param);
                     }
                 }
                 ds.flush();
@@ -398,8 +398,8 @@ public class Trade {
                     List<ItemOption> itemOptions = item.getDisplayOptions();
                     ds.writeByte(itemOptions.size());
                     for (ItemOption io : itemOptions) {
-                        ds.writeByte(io.optionTemplate.id);
-                        ds.writeShort(io.param);
+                        ds.writeShort(io.optionTemplate.id);
+                        ds.writeInt(io.param);
                     }
                 }
                 ds.flush();
